Refactored endpoints to get data using both id and slug

+++ app/Action/Enrollment/DeleteEnrollment.php
@@ -3,16 +3,18 @@
 namespace App\Action\Enrollment;
 
 use App\Models\Enrollment;
+use Illuminate\Support\Facades\Auth;
 
 class DeleteEnrollment
 {
-    public function execute($id)
+    public function execute($studentId, $id)
     {
-        $enrollment = Enrollment::find($id);
+        if (Auth::id() == $studentId) {
+            $enrollment = Enrollment::find($id);
 
-        if(!empty($enrollment))
-        {
-            return $enrollment->delete();
+            if (!empty($enrollment)) {
+                return $enrollment->delete();
+            }
         }
 
         return false;
